DeepL API is now called correctly - Issue #4

+++ app/src/main/java/com/anthony/deepl/model/TranslationResponse.java
@@ -83,7 +83,7 @@ class TranslationResponseResult {
 class TranslationResponseTranslations {
 
     @SerializedName("timeAfterPreprocessing")
-    private int mAfterPreprocessingTime;
+    private Double mAfterPreprocessingTime;
 
     @SerializedName("timeReceivedFromEndpoint")
     private String mReceivedFromEndpointTime;
